# gateway/api/swagger.yaml
basePath: /
definitions:
...
  httpx.ErrorResponse:
    properties:
      error:
        example: access denied
        type: string
    type: object
...
  project.CreateProjectRequest:
    properties:
      actual_deadline:
        type: string
      actual_start_date:
        type: string
      contract_number:
        type: string
      customer_name:
        type: string
      description:
        type: string
      manager_id:
        type: string
      name:
        type: string
      planned_budget:
        type: string
      planned_deadline:
        type: string
      planned_start_date:
        type: string
      project_code:
        type: string
      status:
        type: string
    type: object
...
  project.ListProjectsResponse:
    properties:
      projects:
        items:
          $ref: '#/definitions/project.ProjectResponse'
        type: array
    type: object
...
  project.ProjectResponse:
    properties:
      actual_deadline:
        type: string
      actual_start_date:
        type: string
      contract_number:
        type: string
      created_at:
        type: string
      customer_name:
        type: string
      description:
        type: string
      id:
        type: string
      manager_id:
        type: string
      name:
        type: string
      planned_budget:
        type: string
      planned_deadline:
        type: string
      planned_start_date:
        type: string
      project_code:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  project.UpdateProjectRequest:
    properties:
      actual_deadline:
        type: string
      actual_start_date:
        type: string
      contract_number:
        type: string
      customer_name:
        type: string
      description:
        type: string
      manager_id:
        type: string
      name:
        type: string
      planned_budget:
        type: string
      planned_deadline:
        type: string
      planned_start_date:
        type: string
      project_code:
        type: string
      status:
        type: string
    type: object
...
host: localhost:28080
info:
  contact: {}
  description: Public REST API exposed by the gateway. Internal microservices communicate
    only over gRPC and are intentionally omitted from this specification.
  title: Project Management Gateway API
  version: "1.0"
paths:
...
  /api/projects:
    get:
      description: Returns projects visible to the caller. Administrators may use
        the participant filter, regular users see only their own related projects.
      parameters:
      - description: Participant user ID filter
        in: query
        name: participant_user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ListProjectsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Creates a project. Administrators may assign any manager, regular
        users may create projects only for themselves.
      parameters:
      - description: Project payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create project
      tags:
      - projects
  /api/projects/{id}:
    delete:
      description: Deletes a project. Accessible to administrators and the current
        project manager.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - projects
    get:
      description: Returns a project by ID. Accessible to administrators, the project
        manager, and active project members.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Updates a project. Accessible to administrators and the current
        project manager.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project.UpdateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/project.ProjectResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/httpx.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - projects
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: JWT access token. Use the `Bearer {token}` format.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"